+++ src/texture.rs
@@ -6,7 +6,7 @@ use image::{DynamicImage, GenericImageView, Pixel};
 use std::{fmt::Debug, sync::Arc};
 
 pub trait Texture: Sync + Send + Debug {
-    fn value(&self, u: f64, v: f64, p: Point3) -> Color;
+    fn value(&self, u: f32, v: f32, p: Point3) -> Color;
 }
 
 #[derive(Debug)]
@@ -15,7 +15,7 @@ pub struct SolidColor {
 }
 
 impl Texture for SolidColor {
-    fn value(&self, _u: f64, _v: f64, _p: Point3) -> Color {
+    fn value(&self, _u: f32, _v: f32, _p: Point3) -> Color {
         self.color_value
     }
 }
@@ -36,7 +36,7 @@ impl CheckerTexture {
 }
 
 impl Texture for CheckerTexture {
-    fn value(&self, u: f64, v: f64, p: Point3) -> Color {
+    fn value(&self, u: f32, v: f32, p: Point3) -> Color {
         let sines = (10.0 * p.x()).sin() * (10.0 * p.y()).sin() * (10.0 * p.z()).sin();
         if sines < 0.0 {
             self.odd.value(u, v, p)
@@ -49,11 +49,11 @@ impl Texture for CheckerTexture {
 #[derive(Debug)]
 pub struct NoiseTexture {
     noise: Perlin,
-    scale: f64,
+    scale: f32,
 }
 
 impl NoiseTexture {
-    pub fn new(scale: f64) -> Arc<Self> {
+    pub fn new(scale: f32) -> Arc<Self> {
         Arc::new(Self {
             noise: Perlin::new(),
             scale,
@@ -63,7 +63,7 @@ impl NoiseTexture {
 
 impl Texture for NoiseTexture {
     #[inline]
-    fn value(&self, _u: f64, _v: f64, p: Point3) -> Color {
+    fn value(&self, _u: f32, _v: f32, p: Point3) -> Color {
         rgb!(1.0, 1.0, 1.0)
             * 0.5
             * (1.0 + (self.scale * p.z() + 10.0 * self.noise.turb(p, 7)).sin())
@@ -86,13 +86,13 @@ impl ImageTexture {
 }
 
 impl Texture for ImageTexture {
-    fn value(&self, u: f64, v: f64, _p: Point3) -> Color {
+    fn value(&self, u: f32, v: f32, _p: Point3) -> Color {
         let (width, height) = self.image.dimensions();
         let u = u.max(0.0).min(1.0);
         let v = 1.0 - v.max(0.0).min(1.0);
 
-        let mut i = (u * width as f64) as u32;
-        let mut j = (v * height as f64) as u32;
+        let mut i = (u * width as f32) as u32;
+        let mut j = (v * height as f32) as u32;
 
         if i >= width {
             i = width - 1;
@@ -102,12 +102,12 @@ impl Texture for ImageTexture {
         }
 
         let color_scale = 1.0 / 255.0;
-        let pixel: Vec<f64> = self
+        let pixel: Vec<f32> = self
             .image
             .get_pixel(i, j)
             .channels()
             .iter()
-            .map(|&x| color_scale * x as f64)
+            .map(|&x| color_scale * x as f32)
             .collect();
 
         Color::new(pixel[0], pixel[1], pixel[2])
